apps/backend/collection-service/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
aiofiles>=24.1.0
h5py>=3.11.0
numpy>=1.26.0
aio-pika>=9.4.0
redis>=5.0.0
python-multipart>=0.0.9
pyjwt>=2.8.0
httpx>=0.27.0
python-dotenv>=1.0.0
# ROS 2 bridge (optional — install in ROS 2 env or skip for dev mode)
# rclpy  # provided by ROS 2 installation
mcap>=1.1.0
mcap-ros2-support>=0.5.0
# Policy-server relay
grpcio>=1.64.0
pyzmq>=26.0.0
# Visualisation
rerun-sdk>=0.18.0
